Moved set frame title to InstallerFrame to be set after the language was selected
